pyproject.toml
[project]
name = "vectrify"
version = "0.1.0"
description = "Vectorizes raster images (PNG/JPG) using a mix of LLMs and NSGA-II multi-objective optimization. Outputs SVG and other vector formats."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [
    { name = "Rasmus Ros", email = "rasmus.s.ros@gmail.com" },
]
keywords = [
    "svg",
    "vectorization",
    "raster-to-vector",
    "llm",
    "openai",
    "anthropic",
    "gemini",
    "graphviz",
    "typst",
    "nsga",
    "evolutionary",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.21.0",
    "google-genai>=1.68.0",
    "cairosvg>=2.7.0",
    "pillow>=10.0.0",
    "numpy>=2.2.6",
    "rich>=14.3.3",
    "tqdm>=4.67.3",
]

[project.optional-dependencies]
vision = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
]
graphviz = [
    "graphviz>=0.21",
]
typst = [
    "typst>=0.11.0",
]
all = [
    "vectrify[vision,graphviz,typst]",
]
dev = [
    "pytest",
    "pytest-xdist",
    "ruff",
    "pyrefly",
    "py-spy",
    "matplotlib>=3.10.8",
]

[project.urls]
Homepage = "https://github.com/rasros/vectrify"
Repository = "https://github.com/rasros/vectrify"
Issues = "https://github.com/rasros/vectrify/issues"

[project.scripts]
vectrify = "vectrify.main:main"

[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["vectrify.tests*"]

[tool.pytest.ini_options]
addopts = "-m 'not llm'"
testpaths = ["src"]
markers = [
    "llm: marks tests that require live LLM API calls",
]

[tool.pyrefly]
project-includes = ["src/**/*.py"]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu126" }]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "A",
    "C4",
    "PT",
    "RET",
    "SIM",
    "PTH",
    "PIE",
    "ARG",
    "RUF",
    "TID",
]
ignore = []

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
